move model into separate file
